Move legacy type and slot tests out of test_cpy_compat and into their respective test files.
