+++ test/test_hpytype_legacy.py
@@ -1,5 +1,6 @@
 """ HPyType tests on legacy types. """
 
+from .support import HPyTest
 from .test_hpytype import PointTemplate, TestType as _TestType
 
 
@@ -26,3 +27,51 @@ class LegacyPointTemplate(PointTemplate):
 class TestLegacyType(_TestType):
 
     ExtensionTemplate = LegacyPointTemplate
+
+
+class TestCustomLegacyFeatures(HPyTest):
+
+    def test_legacy_methods(self):
+        mod = self.make_module("""
+            #include <Python.h>
+
+            static PyObject *f(PyObject *self, PyObject *args)
+            {
+                return PyLong_FromLong(1234);
+            }
+            static PyObject *g(PyObject *self, PyObject *arg)
+            {
+                long x = PyLong_AsLong(arg);
+                return PyLong_FromLong(x * 2);
+            }
+            static PyObject *h(PyObject *self, PyObject *args)
+            {
+                long a, b, c;
+                if (!PyArg_ParseTuple(args, "lll", &a, &b, &c))
+                    return NULL;
+                return PyLong_FromLong(100*a + 10*b + c);
+            }
+            static PyObject *k(PyObject *self, PyObject *args, PyObject *kwargs)
+            {
+                static char *kwlist[] = { "a", "b", "c", NULL };
+                long a, b, c;
+                if (!PyArg_ParseTupleAndKeywords(args, kwargs, "lll", kwlist, &a, &b, &c))
+                    return NULL;
+                return PyLong_FromLong(100*a + 10*b + c);
+            }
+
+            static PyMethodDef my_legacy_methods[] = {
+                {"f", (PyCFunction)f, METH_NOARGS},
+                {"g", (PyCFunction)g, METH_O},
+                {"h", (PyCFunction)h, METH_VARARGS},
+                {"k", (PyCFunction)k, METH_VARARGS | METH_KEYWORDS},
+                {NULL}
+            };
+
+            @EXPORT_LEGACY(my_legacy_methods)
+            @INIT
+        """)
+        assert mod.f() == 1234
+        assert mod.g(45) == 90
+        assert mod.h(4, 5, 6) == 456
+        assert mod.k(c=6, b=5, a=4) == 456
